Started Feed page
Created a master page for the site and started the feed page.
