+++ feed.aspx
@@ -1,0 +1,9 @@
+﻿<%@ Page Language="C#" AutoEventWireup="true" CodeFile="feed.aspx.cs" Inherits="feed" MasterPageFile="~/Master_Pages/main.master" %>
+
+<asp:Content ContentPlaceHolderID="HeaderAreacp" runat="server">
+
+</asp:Content>
+
+<asp:Content ContentPlaceHolderID="MainContentcp" runat="server">
+    <asp:Panel ID="FeedContentpnl" runat="server"></asp:Panel>
+</asp:Content>
